Убрал шаблоны, с ними сложнее. Нашёл, кажется, как исправить ошибку, но пока сложно. Исправил целевую папку для программы

+++ task-cmake/C/main.cpp
@@ -5,7 +5,7 @@ int main() {
   size_t n;
   std::cin >> n;
 
-  std::vector<int64_t> arr(n);
+  std::vector<int> arr(n);
   for (size_t i = 0; i < n; i++) {
     std::cin >> arr[i];
   }
